refactor: (#152) 댓글 컴포넌트들을 comment 폴더로 이동 댓글에 관련된 컴포넌트 수가 많아져서 따로 분리하였음
